==> stage-I-IDH-NLP-ClassificationModel/06.Performace_table.ipynb <==
import pandas as pd, numpy as np, glob, os

# calc fulltext and sentence performace
model_name_s = "ST-DBMCv1" #ST-DBMCv1, mUSE
ths = 0.5
fulltext_mUSE = pd.read_csv(f'./other_experiment/ACM_mUSE_performace_table_predict_test_sample500_intra+peri-ths{ths}_fulltext.csv')
fulltext_mSBERT = pd.read_csv(f'./other_experiment/ACM_ST-DBMCv1_performace_table_predict_test_sample500_intra+peri-ths{ths}_fulltext.csv')
sentence_mUSE = pd.read_csv(f'./ACM_mUSE_performace_table_predict_test_sample500_intra+peri-ths{ths}.csv')
sentence_mSBERT = pd.read_csv(f'./ACM_ST-DBMCv1_performace_table_predict_test_sample500_intra+peri-ths{ths}.csv')
new_dict = {'mSBERT':sentence_mSBERT['IDH Judge'], 'mUSE': sentence_mUSE['IDH Judge'],
            'mSBERT-F':fulltext_mSBERT['IDH Judge'], 'mUSE-F': fulltext_mUSE['IDH Judge'],}
performance_fulltext_vs_sentence = (pd.DataFrame(new_dict).transpose())
performance_fulltext_vs_sentence[[0,1,2,3,4]] = performance_fulltext_vs_sentence[[0,1,2,3,4]].astype(float).round(3)
performance_fulltext_vs_sentence = performance_fulltext_vs_sentence.rename(columns={0:'Accruacy',1:'Precision [macro]',2:'Recall [macro]',
                                                                                    3:'Specificity',4:'F1-score [macro]',5:'TP / TN', 6:'FP / FN'})
performance_fulltext_vs_sentence.round(3).reset_index().rename(columns={'index':'Model [Thresholds=0.5]'})
